% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDQutils.r
\name{PDQutils-NEWS}
\alias{PDQutils-NEWS}
\title{News for package 'PDQutils':

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\PDQutils}{\CRANpkg{PDQutils}}}
\description{
News for package 'PDQutils':

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\PDQutils}{\CRANpkg{PDQutils}}
}
\section{\PDQutils{} Version 0.1.1 (2015-02-26) }{

\itemize{
\item Edgeworth expansions.
}
}

\section{\PDQutils{} Initial Version 0.1.0 (2015-02-14) }{

\itemize{
\item first CRAN release.
}
}

