% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDQutils.r
\name{PDQutils-NEWS}
\alias{PDQutils-NEWS}
\title{News for package \sQuote{PDQutils}:}
\description{
News for package \sQuote{PDQutils}.

\newcommand{\pkg}{#1}
\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\PDQutils}{\CRANpkg{PDQutils}}
}
\section{\PDQutils{} Version 0.1.6 (2017-03-18) }{

\itemize{
\item Package maintenance--no new features.
\item move github figures to location CRAN understands.
}
}

\section{\PDQutils{} Version 0.1.5 (2016-09-18) }{

\itemize{
\item Package maintenance--no new features.
\item Remove errant files from test directory.
}
}

\section{\PDQutils{} Version 0.1.4 (2016-03-03) }{

\itemize{
\item Package maintenance--no new features.
\item Incompatibilities in vignette with ggplot2 release.
}
}

\section{\PDQutils{} Version 0.1.3 (2016-01-04) }{

\itemize{
\item Package maintenance--no new features.
}
}

\section{\PDQutils{} Version 0.1.2 (2015-06-15) }{

\itemize{
\item Generalized Gram Charlier expansions.
\item bugfixes.
}
}

\section{\PDQutils{} Version 0.1.1 (2015-02-26) }{

\itemize{
\item Edgeworth expansions.
}
}

\section{\PDQutils{} Initial Version 0.1.0 (2015-02-14) }{

\itemize{
\item first CRAN release.
}
}

